+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.2.0] - 2026-08-12
+
+### Added
+- Added `RepositoryFactory` for exposing custom Cycle repositories as container services using their entity-schema mapping.
+- Added repository service documentation with read and persist repository examples.
+
 ## [4.1.1] - 2026-08-10
 
 ### Changed